+++ src/models/Abstract.php
@@ -75,7 +75,8 @@ class Shopgate_Model_Abstract extends ShopgateObject
         switch (substr($method, 0, 3)) {
             case 'get':
                 $key  = $this->underscore(substr($method, 3));
-                $data = $this->getData($key,
+                $data = $this->getData(
+                    $key,
                     isset($args[0])
                         ? $args[0]
                         : null
@@ -84,7 +85,8 @@ class Shopgate_Model_Abstract extends ShopgateObject
                 return $data;
             case 'set':
                 $key    = $this->underscore(substr($method, 3));
-                $result = $this->setData($key,
+                $result = $this->setData(
+                    $key,
                     isset($args[0])
                         ? $args[0]
                         : null