+++ src/models/catalog/Category.php
@@ -118,7 +118,8 @@ class Shopgate_Model_Catalog_Category extends Shopgate_Model_AbstractExport
         $categoryNode = $itemNode->addChild('category');
         $categoryNode->addAttribute('uid', $this->getUid());
         $categoryNode->addAttribute('sort_order', (int)$this->getSortOrder());
-        $categoryNode->addAttribute('parent_uid',
+        $categoryNode->addAttribute(
+            'parent_uid',
             $this->getParentUid()
                 ? $this->getParentUid()
                 : null