+++ tests/unit/helper/error_handling/ErrorHandlerTest.php
@@ -19,12 +19,15 @@
  * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
-class Shopgate_Helper_Error_Handling_ErrorHandlerTest extends PHPUnit_Framework_TestCase
+
+namespace shopgate\cart_integration_sdk\tests\unit\error_handling;
+
+class ErrorHandlerTest extends \PHPUnit_Framework_TestCase
 {
-    /** @var PHPUnit_Framework_MockObject_MockObject|Shopgate_Helper_Logging_Stack_Trace_GeneratorInterface */
+    /** @var \PHPUnit_Framework_MockObject_MockObject|\Shopgate_Helper_Logging_Stack_Trace_GeneratorInterface */
     protected $stackTraceGenerator;
 
-    /** @var PHPUnit_Framework_MockObject_MockObject|Shopgate_Helper_Logging_Strategy_LoggingInterface */
+    /** @var \PHPUnit_Framework_MockObject_MockObject|\Shopgate_Helper_Logging_Strategy_LoggingInterface */
     protected $logging;
 
     public function setUp()
@@ -45,20 +48,20 @@ class Shopgate_Helper_Error_Handling_ErrorHandlerTest extends PHPUnit_Framework_
         $this->stackTraceGenerator
             ->expects($this->once())
             ->method('generate')
-            ->with(new PHPUnit_Framework_Constraint_Exception('Exception'))
+            ->with(new \PHPUnit_Framework_Constraint_Exception('Exception'))
             ->willReturn($stackTrace);
 
         $this->logging
             ->expects($this->once())
             ->method('log')
             ->with(
-                new PHPUnit_Framework_Constraint_IsAnything(), // message
-                Shopgate_Helper_Logging_Strategy_LoggingInterface::LOGTYPE_ERROR,
+                new \PHPUnit_Framework_Constraint_IsAnything(), // message
+                \Shopgate_Helper_Logging_Strategy_LoggingInterface::LOGTYPE_ERROR,
                 $stackTrace
             )
             ->willReturn(true);
 
-        $SUT = new Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
+        $SUT = new \Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
         $SUT->handle(123, 'a message', '/var/www/failingscript.php', 100, array());
     }
 
@@ -72,7 +75,7 @@ class Shopgate_Helper_Error_Handling_ErrorHandlerTest extends PHPUnit_Framework_
             ->expects($this->never())
             ->method('log');
 
-        $SUT = new Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
+        $SUT = new \Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
         $SUT->handle(0, 'a message', '/var/www/failingscript.php', 100, array());
     }
 
@@ -82,15 +85,15 @@ class Shopgate_Helper_Error_Handling_ErrorHandlerTest extends PHPUnit_Framework_
         // "If the function returns FALSE then the normal error handler continues."
 
         // internal error handler should be used by default
-        $SUT = new Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
+        $SUT = new \Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging);
         $this->assertFalse($SUT->handle(123, 'a message', '/var/www/failingscript.php', 100, array()));
 
         // internal error handler explicitly used
-        $SUT = new Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging, false);
+        $SUT = new \Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging, false);
         $this->assertFalse($SUT->handle(123, 'a message', '/var/www/failingscript.php', 100, array()));
 
         // internal error handler disabled
-        $SUT = new Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging, true);
+        $SUT = new \Shopgate_Helper_Error_Handling_ErrorHandler($this->stackTraceGenerator, $this->logging, true);
         $this->assertTrue($SUT->handle(123, 'a message', '/var/www/failingscript.php', 100, array()));
     }
 }