+++ tests/unit/helper/logging/ObfuscatorTest.php
@@ -19,14 +19,17 @@
  * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
-class Shopgate_Helper_Logging_ObfuscatorTest extends PHPUnit_Framework_TestCase
+
+namespace shopgate\cart_integration_sdk\tests\unit\logging;
+
+class ObfuscatorTest extends \PHPUnit_Framework_TestCase
 {
-    /** @var Shopgate_Helper_Logging_Obfuscator */
+    /** @var \Shopgate_Helper_Logging_Obfuscator */
     private $obfuscator;
 
     public function setUp()
     {
-        $this->obfuscator = new Shopgate_Helper_Logging_Obfuscator();
+        $this->obfuscator = new \Shopgate_Helper_Logging_Obfuscator();
     }
 
     public function testAddObfuscationFields()