+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
 - modified _.htaccess_ file to support Apache 2.4 module _mod_authz_core_
+- introduced namespaces in folder tests
+
+### Added
+- white list for jobs in method cron
 
 ## [2.9.69] - 2017-09-13
 ### Added